moved the createplaceidapiusingpostrequest class file under utilities folder and updated the postrequestusing file test script under test script folder
