+++ src/main/java/utilities/CreatePlaceIdAPIUsingPostRequest.java
@@ -1,4 +1,4 @@
-package testscript;
+package utilities;
 
 import io.restassured.response.Response;
 import jsonRequestAndResponse.JsonPostRequest;
