+++ test-output/Default suite/Default test.html
@@ -55,11 +55,11 @@ function toggleAllBoxes() {
 <body>
 <h2 align='center'>Default test</h2><table border='1' align="center">
 <tr>
-<td>Tests passed/Failed/Skipped:</td><td>1/0/0</td>
+<td>Tests passed/Failed/Skipped:</td><td>2/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Thu May 26 15:27:01 EDT 2022</td>
+<td>Started on:</td><td>Thu May 26 15:31:37 EDT 2022</td>
 </tr>
-<tr><td>Total time:</td><td>3 seconds (3368 ms)</td>
+<tr><td>Total time:</td><td>7 seconds (7581 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -75,10 +75,15 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
+<td title='testscript.GetRequest.getRequest()'><b>getRequest</b><br>Test class: testscript.GetRequest</td>
+<td></td>
+<td>6</td>
+<td>testscript.GetRequest@6c9f5c0d</td></tr>
+<tr>
 <td title='testscript.PostRequestUsingFile.postRequestUsingFile()'><b>postRequestUsingFile</b><br>Test class: testscript.PostRequestUsingFile</td>
 <td></td>
-<td>3</td>
-<td>testscript.PostRequestUsingFile@2698dc7</td></tr>
+<td>1</td>
+<td>testscript.PostRequestUsingFile@276438c9</td></tr>
 </table><p>
 </body>
 </html>
