+++ test-output/Default suite/Default test.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite hostname="DESKTOP-6E7JKDN" name="Default test" tests="1" failures="0" timestamp="26 May 2022 19:27:05 GMT" time="3.368" errors="0">
-  <testcase name="postRequestUsingFile" time="3.354" classname="testscript.PostRequestUsingFile"/>
+<testsuite hostname="DESKTOP-6E7JKDN" name="Default test" tests="2" failures="0" timestamp="26 May 2022 19:31:45 GMT" time="7.581" errors="0">
+  <testcase name="getRequest" time="6.108" classname="testscript.GetRequest"/>
+  <testcase name="postRequestUsingFile" time="1.458" classname="testscript.PostRequestUsingFile"/>
 </testsuite> <!-- Default test -->
