+++ test-output/Suite/Test.html
@@ -55,11 +55,11 @@ function toggleAllBoxes() {
 <body>
 <h2 align='center'>Test</h2><table border='1' align="center">
 <tr>
-<td>Tests passed/Failed/Skipped:</td><td>1/0/0</td>
+<td>Tests passed/Failed/Skipped:</td><td>2/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Thu May 26 15:18:45 EDT 2022</td>
+<td>Started on:</td><td>Thu May 26 15:29:38 EDT 2022</td>
 </tr>
-<tr><td>Total time:</td><td>5 seconds (5352 ms)</td>
+<tr><td>Total time:</td><td>7 seconds (7399 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -78,7 +78,12 @@ function toggleAllBoxes() {
 <td title='testscript.GetRequest.getRequest()'><b>getRequest</b><br>Test class: testscript.GetRequest</td>
 <td></td>
 <td>5</td>
-<td>testscript.GetRequest@2698dc7</td></tr>
+<td>testscript.GetRequest@43d7741f</td></tr>
+<tr>
+<td title='testscript.PostRequestUsingFile.postRequestUsingFile()'><b>postRequestUsingFile</b><br>Test class: testscript.PostRequestUsingFile</td>
+<td></td>
+<td>1</td>
+<td>testscript.PostRequestUsingFile@60438a68</td></tr>
 </table><p>
 </body>
 </html>
