+++ test-output/Suite/Test.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite hostname="DESKTOP-6E7JKDN" name="Test" tests="1" failures="0" timestamp="26 May 2022 19:18:51 GMT" time="5.352" errors="0">
-  <testcase name="getRequest" time="5.337" classname="testscript.GetRequest"/>
+<testsuite hostname="DESKTOP-6E7JKDN" name="Test" tests="2" failures="0" timestamp="26 May 2022 19:29:45 GMT" time="7.399" errors="0">
+  <testcase name="getRequest" time="5.917" classname="testscript.GetRequest"/>
+  <testcase name="postRequestUsingFile" time="1.467" classname="testscript.PostRequestUsingFile"/>
 </testsuite> <!-- Test -->
