+++ test-output/emailable-report.html
@@ -31,14 +31,17 @@ div.shootout {padding:2em;border:3px #4854A8 solid}
 <body>
 <table cellspacing="0" cellpadding="0" class="result">
 <tr><th>Test</th><th class="numi">Methods<br/>Passed</th><th class="numi">Scenarios<br/>Passed</th><th class="numi"># skipped</th><th class="numi"># failed</th><th class="numi">Total<br/>Time</th><th class="numi">Included<br/>Groups</th><th class="numi">Excluded<br/>Groups</th></tr>
-<tr><td style="text-align:left;padding-right:2em">Default test</td><td class="numi">1</td><td class="numi">1</td><td class="numi">0</td><td class="numi">0</td><td class="numi">3.4 seconds</td><td class="numi"></td><td class="numi"></td></tr>
+<tr><td style="text-align:left;padding-right:2em">Default test</td><td class="numi">2</td><td class="numi">2</td><td class="numi">0</td><td class="numi">0</td><td class="numi">7.6 seconds</td><td class="numi"></td><td class="numi"></td></tr>
 </table>
 <table cellspacing="0" cellpadding="0" class="passed" id="summary">
 <tr><th>Class</th><th>Method</th><th># of<br/>Scenarios</th><th>Start</th><th>Time<br/>(ms)</th></tr>
 <tr><th colspan="4">Default test &#8212; passed</th></tr>
-<tr class="passedodd"><td rowspan="1">testscript.PostRequestUsingFile</td><td><a href="#m1"><b>postRequestUsingFile</b>  </a></td><td class="numi">1</td><td>1653593221633</td><td class="numi">3354</td></tr>
+<tr class="passedodd"><td rowspan="1">testscript.GetRequest</td><td><a href="#m1"><b>getRequest</b>  </a></td><td class="numi">1</td><td>1653593497355</td><td class="numi">6108</td></tr>
+<tr class="passedeven"><td rowspan="1">testscript.PostRequestUsingFile</td><td><a href="#m2"><b>postRequestUsingFile</b>  </a></td><td class="numi">1</td><td>1653593503466</td><td class="numi">1458</td></tr>
 </table>
 <h1>Default test</h1>
-<h2 id="m1">testscript.PostRequestUsingFile:postRequestUsingFile</h2>
+<h2 id="m1">testscript.GetRequest:getRequest</h2>
+<p class="totop"><a href="#summary">back to summary</a></p>
+<h2 id="m2">testscript.PostRequestUsingFile:postRequestUsingFile</h2>
 <p class="totop"><a href="#summary">back to summary</a></p>
 </body></html>
