+++ test-output/junitreports/TEST-testscript.GetRequest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="DESKTOP-6E7JKDN" name="testscript.GetRequest" tests="1" failures="0" timestamp="26 May 2022 19:18:51 GMT" time="5.337" errors="0">
-  <testcase name="getRequest" time="5.337" classname="testscript.GetRequest"/>
+<testsuite hostname="DESKTOP-6E7JKDN" name="testscript.GetRequest" tests="1" failures="0" timestamp="26 May 2022 19:31:45 GMT" time="6.108" errors="0">
+  <testcase name="getRequest" time="6.108" classname="testscript.GetRequest"/>
 </testsuite> <!-- testscript.GetRequest -->
