+++ test-output/junitreports/TEST-testscript.PostRequestUsingFile.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="DESKTOP-6E7JKDN" name="testscript.PostRequestUsingFile" tests="1" failures="0" timestamp="26 May 2022 19:27:05 GMT" time="3.354" errors="0">
-  <testcase name="postRequestUsingFile" time="3.354" classname="testscript.PostRequestUsingFile"/>
+<testsuite hostname="DESKTOP-6E7JKDN" name="testscript.PostRequestUsingFile" tests="1" failures="0" timestamp="26 May 2022 19:31:45 GMT" time="1.458" errors="0">
+  <testcase name="postRequestUsingFile" time="1.458" classname="testscript.PostRequestUsingFile"/>
 </testsuite> <!-- testscript.PostRequestUsingFile -->
