+++ test-output/old/Default suite/classes.html
@@ -4,6 +4,28 @@
 <th>Method name</th>
 <th>Groups</th>
 </tr><tr>
+<td>testscript.GetRequest</td>
+<td>&nbsp;</td><td>&nbsp;</td></tr>
+<tr>
+<td align='center' colspan='3'>@Test</td>
+</tr>
+<tr>
+<td>&nbsp;</td>
+<td>getRequest</td>
+<td>&nbsp;</td></tr>
+<tr>
+<td align='center' colspan='3'>@BeforeClass</td>
+</tr>
+<tr>
+<td align='center' colspan='3'>@BeforeMethod</td>
+</tr>
+<tr>
+<td align='center' colspan='3'>@AfterMethod</td>
+</tr>
+<tr>
+<td align='center' colspan='3'>@AfterClass</td>
+</tr>
+<tr>
 <td>testscript.PostRequestUsingFile</td>
 <td>&nbsp;</td><td>&nbsp;</td></tr>
 <tr>
