+++ test-output/old/Default suite/methods-alphabetical.html
@@ -1,6 +1,8 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Default suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="a58fbd">  <td>22/05/26 15:27:01</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="PostRequestUsingFile.postRequestUsingFile()[pri:0, instance:testscript.PostRequestUsingFile@2698dc7]">postRequestUsingFile</td> 
-  <td>main@1143371233</td>   <td></td> </tr>
+<tr bgcolor="a58fbd">  <td>22/05/26 15:31:37</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="GetRequest.getRequest()[pri:0, instance:testscript.GetRequest@6c9f5c0d]">getRequest</td> 
+  <td>main@37380050</td>   <td></td> </tr>
+<tr bgcolor="d4e2ba">  <td>22/05/26 15:31:43</td>   <td>6111</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="PostRequestUsingFile.postRequestUsingFile()[pri:0, instance:testscript.PostRequestUsingFile@276438c9]">postRequestUsingFile</td> 
+  <td>main@37380050</td>   <td></td> </tr>
 </table>
