+++ test-output/old/Suite/methods-alphabetical.html
@@ -1,6 +1,8 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="a58fbd">  <td>22/05/26 15:18:45</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="GetRequest.getRequest()[pri:0, instance:testscript.GetRequest@2698dc7]">getRequest</td> 
-  <td>main@1161667116</td>   <td></td> </tr>
+<tr bgcolor="a58fbd">  <td>22/05/26 15:29:38</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="GetRequest.getRequest()[pri:0, instance:testscript.GetRequest@43d7741f]">getRequest</td> 
+  <td>main@1143371233</td>   <td></td> </tr>
+<tr bgcolor="9c6b77">  <td>22/05/26 15:29:44</td>   <td>5920</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="PostRequestUsingFile.postRequestUsingFile()[pri:0, instance:testscript.PostRequestUsingFile@60438a68]">postRequestUsingFile</td> 
+  <td>main@1143371233</td>   <td></td> </tr>
 </table>
