+++ test-output/testng-results.xml
@@ -1,13 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<testng-results skipped="0" failed="0" total="1" passed="1">
+<testng-results skipped="0" failed="0" total="2" passed="2">
   <reporter-output>
   </reporter-output>
-  <suite name="Default suite" duration-ms="3368" started-at="2022-05-26T19:27:01Z" finished-at="2022-05-26T19:27:04Z">
+  <suite name="Default suite" duration-ms="7581" started-at="2022-05-26T19:31:37Z" finished-at="2022-05-26T19:31:44Z">
     <groups>
     </groups>
-    <test name="Default test" duration-ms="3368" started-at="2022-05-26T19:27:01Z" finished-at="2022-05-26T19:27:04Z">
+    <test name="Default test" duration-ms="7581" started-at="2022-05-26T19:31:37Z" finished-at="2022-05-26T19:31:44Z">
+      <class name="testscript.GetRequest">
+        <test-method status="PASS" signature="getRequest()[pri:0, instance:testscript.GetRequest@6c9f5c0d]" name="getRequest" duration-ms="6108" started-at="2022-05-26T15:31:37Z" finished-at="2022-05-26T15:31:43Z">
+          <reporter-output>
+          </reporter-output>
+        </test-method> <!-- getRequest -->
+      </class> <!-- testscript.GetRequest -->
       <class name="testscript.PostRequestUsingFile">
-        <test-method status="PASS" signature="postRequestUsingFile()[pri:0, instance:testscript.PostRequestUsingFile@2698dc7]" name="postRequestUsingFile" duration-ms="3354" started-at="2022-05-26T15:27:01Z" finished-at="2022-05-26T15:27:04Z">
+        <test-method status="PASS" signature="postRequestUsingFile()[pri:0, instance:testscript.PostRequestUsingFile@276438c9]" name="postRequestUsingFile" duration-ms="1458" started-at="2022-05-26T15:31:43Z" finished-at="2022-05-26T15:31:44Z">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- postRequestUsingFile -->
